requirements.txt
Django>=5.0,<7
Pillow>=10.0.0
pdf2image>=1.17.0
PyMuPDF>=1.23.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
drf-spectacular
django-environ
openpyxl
cryptography

# Production (Railway / gunicorn)
gunicorn>=22.0
whitenoise>=6.6

# Redis cache + Celery
django-redis>=5.4
celery[redis]>=5.3

# Optional S3-compatible media (USE_S3=1)
django-storages>=1.14
boto3>=1.34
